moved unit tests to separate file
