+++ tests/rest/test_connection.py
@@ -507,54 +507,3 @@ def test_get_process_graph(requests_mock):
     udp = conn.process_graph('evi')
 
     assert udp.process_graph_id == 'evi'
-
-
-def test_describe_process_graph(requests_mock):
-    # TODO: move this method to tests/udp.py
-    requests_mock.get(API_URL, json={"api_version": "1.0.0"})
-    conn = Connection(API_URL)
-
-    # TODO: read it from a file?
-    requests_mock.get(API_URL + "process_graphs/evi", json={
-        "id": "evi",
-        "summary": "Enhanced Vegetation Index",
-        "description": "Computes the Enhanced Vegetation Index (EVI). It is computed with the following formula: `2.5 * (NIR - RED) / (1 + NIR + 6*RED + -7.5*BLUE)`.",
-    })
-
-    udp = conn.process_graph(process_graph_id='evi')
-    details = udp.describe()
-
-    assert details['summary'] == "Enhanced Vegetation Index"
-
-
-def test_update_process_graph(requests_mock):
-    # TODO: move this method to tests/udp.py
-    requests_mock.get(API_URL, json={"api_version": "1.0.0"})
-    conn = Connection(API_URL)
-
-    updated_process_graph = {
-        'description': "Computes the Enhanced Vegetation Index (EVI).",
-        'parameters': [],
-        'process_graph': {
-            'sub': {
-                'process_id': 'add'
-            }
-        }
-    }
-
-    def check_body(request):
-        assert request.json() == updated_process_graph
-        return True
-
-    requests_mock.put(API_URL + "process_graphs/evi", additional_matcher=check_body)
-
-    udp = conn.process_graph(process_graph_id='evi')
-    udp.update(
-        process_graph={
-            'sub': {
-                'process_id': 'add'
-            }
-        },
-        description="Computes the Enhanced Vegetation Index (EVI).",
-        parameters=[]
-    )
